+++ apps/docs/src/app/core/api-files.ts
@@ -398,5 +398,6 @@ export const API_FILES = {
         'TimelineNodeHeaderInfoTextDirective',
         'TimelineNodeActionDirective',
         'TimelineHeaderInfoSubTitleDirective'
-    ]
+    ],
+    scrollbar: ['ScrollbarComponent']
 };