+++ apps/docs/src/app/core/documentation/core-documentation.component.ts
@@ -97,7 +97,8 @@ export class CoreDocumentationComponent extends DocumentationBaseComponent {
             { url: 'core/slider', name: 'Slider' },
             { url: 'core/timeline', name: 'Timeline' },
             { url: 'core/splitter', name: 'Splitter/Responsive Splitter' },
-            { url: 'core/progress-indicator', name: 'Progress Indicator' }
+            { url: 'core/progress-indicator', name: 'Progress Indicator' },
+            { url: 'core/scrollbar', name: 'Scrollbar' }
         ];
 
         this.layouts = [